Yuvish app

+++ app/src/main/java/com/example/yuvish/Models/ArrangedSubmit/Buyurtmalar.kt
@@ -2,7 +2,7 @@ package com.example.yuvish.Models.ArrangedSubmit
 
 data class Buyurtmalar(
     val count: Int,
-    val hajm: Int,
+    val hajm: Double,
     val id: Int,
     val name: String,
     val olchov: String,
