+++ app/src/main/java/com/example/yuvish/Models/ArrangedSubmit/Product.kt
@@ -1,9 +1,9 @@
 package com.example.yuvish.Models.ArrangedSubmit
 
 data class Product(
-    val clean_hajm: Int,
-    val gilam_boyi: Int,
-    val gilam_eni: Int,
+    val clean_hajm: Double,
+    val gilam_boyi: Double,
+    val gilam_eni: Double,
     val id: Int,
     val joy: String,
     val narx: Int,
